backend/app/requirements.txt
# Web Framework
fastapi
uvicorn
pydantic


# Database & Vector Store
beanie
motor  # Async driver for MongoDB, used by Beanie
chromadb
redis

# Asynchronous Tasks
celery

# LangChain & Document Processing
langchain
langchain-community
langchain-text-splitters
unstructured[pdf] # For intelligent PDF parsing
pymupdf           # A dependency for unstructured
sentence-transformers # For creating embeddings

# Configuration & Utilities
pydantic-settings
openai
python-dotenv
python-multipart # For file uploads